Fix lost disk index data

+++ internal/core/src/storage/DiskFileManagerImpl.cpp
@@ -93,7 +93,7 @@ DiskFileManagerImpl::AddFile(const std::string& file) noexcept {
     // Split local data to multi part with specified size
     int slice_num = 0;
     auto remotePrefix = GetRemoteIndexObjectPrefix();
-    for (int offset = 0; offset < fileSize; slice_num++) {
+    for (int64_t offset = 0; offset < fileSize; slice_num++) {
         auto batch_size = std::min(milvus::config::KnowhereGetIndexSliceSize() << 20, int64_t(fileSize) - offset);
 
         auto fieldData = std::make_shared<FieldData>(buf.get() + offset, batch_size);
